# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/monthlyprocess/aggr/MonthlyAggregationEmployeeServiceImpl.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/monthlyprocess/aggr/MonthlyAggregationEmployeeServiceImpl.java
@@ -254,24 +254,56 @@ public class MonthlyAggregationEmployeeServiceImpl implements MonthlyAggregation
 			// 計算結果と同月データ・締めID違い かつ 期間重複データの削除
 			val attendanceTimeOlds = this.attendanceTimeRepository.findByYearMonthOrderByStartYmd(employeeId, yearMonth);
 			for (val oldData : attendanceTimeOlds){
+				val oldClosureId = oldData.getClosureId();
+				val oldClosureDate = oldData.getClosureDate();
+				
 				if (!this.periodCompareEx(oldData.getDatePeriod(), datePeriod)) continue;
 				boolean isTarget = false;
-				if (oldData.getClosureId().value != closureId.value) isTarget = true;
+				if (oldClosureId.value != closureId.value) isTarget = true;
+				if (oldClosureDate.getClosureDay().v() != closureDate.getClosureDay().v()) isTarget = true;
+				if (oldClosureDate.getLastDayOfMonth() != closureDate.getLastDayOfMonth()) isTarget = true;
 				if (!isTarget) continue;
 				this.attendanceTimeRepository.remove(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
-				this.anyItemRepository.removeByMonthlyAndClosure(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
-				this.annLeaRemNumEachMonthRepo.remove(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
-				this.rsvLeaRemNumEachMonthRepo.remove(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
-				this.absLeaRemRepo.remove(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
-				this.monDayoffRemRepo.remove(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
-				this.spcLeaRemRepo.remove(
-						employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
+						employeeId, yearMonth, oldClosureId, oldClosureDate);
+				
+				if (this.affiliationInfoRepository.find(
+						employeeId, yearMonth, oldClosureId, oldClosureDate).isPresent()){
+					this.affiliationInfoRepository.remove(
+							employeeId, yearMonth, oldClosureId, oldClosureDate);
+				}
+				
+				if (this.anyItemRepository.findByMonthlyAndClosure(
+						employeeId, yearMonth, oldClosureId, oldClosureDate).size() > 0){
+					this.anyItemRepository.removeByMonthlyAndClosure(
+							employeeId, yearMonth, oldClosureId, oldClosureDate);
+				}
+				
+				if (this.annLeaRemNumEachMonthRepo.find(
+						employeeId, yearMonth, oldClosureId, oldClosureDate).isPresent()){
+					this.annLeaRemNumEachMonthRepo.remove(
+							employeeId, yearMonth, oldClosureId, oldClosureDate);
+				}
+				
+				if (this.rsvLeaRemNumEachMonthRepo.find(
+						employeeId, yearMonth, oldClosureId, oldClosureDate).isPresent()){
+					this.rsvLeaRemNumEachMonthRepo.remove(
+							employeeId, yearMonth, oldClosureId, oldClosureDate);
+				}
+				
+				if (this.absLeaRemRepo.find(employeeId, yearMonth, oldClosureId, oldClosureDate).isPresent()){
+					this.absLeaRemRepo.remove(
+							employeeId, yearMonth, oldClosureId, oldClosureDate);
+				}
+				
+				if (this.monDayoffRemRepo.find(employeeId, yearMonth, oldClosureId, oldClosureDate).isPresent()){
+					this.monDayoffRemRepo.remove(
+							employeeId, yearMonth, oldClosureId, oldClosureDate);
+				}
+				
+				if (this.spcLeaRemRepo.find(employeeId, yearMonth, oldClosureId, oldClosureDate).size() > 0){
+					this.spcLeaRemRepo.remove(
+							employeeId, yearMonth, oldData.getClosureId(), oldData.getClosureDate());
+				}
 			}
 			
 			// 登録する
